pyproject.toml
[project]
name = "trade_exec_gym"
version = "0.1.0"
description = "HuggingFace x Meta OpenEnv Hackathon - Smart Order Router"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openenv-core>=0.1.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.30.6",
    "pydantic>=2.9.2",
    "fastmcp",
    "numpy",
    "scipy",
    "pytest",
    "requests",
    "httpx",
    "openai",
    "gradio",
    "matplotlib",
    "pandas",
    "torch>=2.1.0",
    "stable-baselines3>=2.1.0",
    "tensorboard",
]

[project.scripts]
server = "server.app:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["server", "env", "baselines", "training", "tasks"]

[tool.pytest.ini_options]
pythonpath = ["."]
